Checks if there is a newer version of R

Fetches the latest (not development!) R version and compares it with your currently installed R version (the version of the R session from which you are running this function).

check.for.updates.R(
  notify_user = TRUE,
  GUI = TRUE,
  page_with_download_url = "https://cran.rstudio.com/bin/windows/base/",
  pat = "R-[0-9.]+.+-win\\.exe"
)

Arguments

notify_user
if to print to you (the user) what is the latest version and what version you are currently using.

GUI
a logical indicating whether a graphics menu should be used if available. If TRUE, and on Windows, it will use winDialog, otherwise it will use cat.

page_with_download_url
the URL of the page from which R can be downloaded.

pat
pattern to search for when looking for a newer R version

Value

TRUE/FALSE - if there is a newer version of R to install or not.

Examples

if (FALSE) {
check.for.updates.R() 
# Possible output:
#  There is a newer version of R for you to download!
#  You are using R version:  2.15.0 
#  And the latest R version is:  2.15.3 
#  [1] TRUE
}