Read R Code from a File in an https URL

source.https causes R to accept its input from a File in an https URL. Input is read and parsed from that file until the end of the file is reached, then the parsed expressions are evaluated sequentially in the chosen environment.

source.https(URL, ..., remove_r_file = T)

Arguments

URL
the URL of the .r file to download and source.

...
parameters to pass to source

remove_r_file
if to remove the .r file after it was sourced.

Value

Nothing.

Details

"The easiest way to use Git on Windows." (at least so they say...)

References

Other solutions to the source.https problem:

Using RCurl

devtools::source_url

A relevant (OLD) discussion: http://stackoverflow.com/questions/7715723/sourcing-r-script-over-https

See also

Examples

# NOT RUN {
source.https("https://raw.github.com/talgalili/installr/master/R/install.r")
# }