Uninstall an R version

Choose an R version to uninstall via a menubar. By default, the function allows the user to pick an R version to uninstall from a list. Also, the function can be called with using "r_version", where multiple R versions can be supplied and all will be uninstalled.

uninstall.R(r_version, GUI = TRUE)

Arguments

r_version
a character vector for R versions to uninstall (the format is of the style: "2.15.3"). default is empty - resulting in a prompt message asking the user what to do.

GUI
If asking the user which R version to uninstall, should the GUI be used? (default is TRUE)

Value

the output of system running the uninstaller

See also

Examples

if (FALSE) {
uninstall.R() # choose an R version to uninstall
uninstall.R("2.15.3") # will uninstall R 2.15.3
uninstall.R(c("2.15.3", "2.14.0")) # will uninstall two R versions (if both exists)
uninstall.R("10.10.0") # would pop up the menu options (until R 10.10.0 will be released :D )
}