uninstalls (removes) Installed Packages

A wrapper for remove.packages. Useful since it also works if the package is currently loaded into the workspace.

uninstall.packages(pkgs, lib, warning = TRUE, ...)

Arguments

pkgs
a character vector with the names of the packages to be removed.

lib
a character vector giving the library directories to remove the packages from. If missing, defaults to the first element in .libPaths.

warning
boolean (TRUE), should a message be printed in various cases.

...
currently ignored.

Value

Invisible NULL

See also

Examples

if (FALSE) {
install.packages(c("reshape", "plyr"))
require(plyr)
uninstall.packages(c("reshape", "plyr"))
install.packages(c("reshape", "plyr"))
}